--- a/lib/prometheus/client/label_set_validator.rb
+++ b/lib/prometheus/client/label_set_validator.rb
@@ -33,15 +33,17 @@ module Prometheus
       end
 
       def validate_labelset!(labelset)
-        validate_symbols!(labelset)
-
-        unless keys_match?(labelset)
-          raise InvalidLabelSetError, "labels must have the same signature " \
-                                      "(keys given: #{labelset.keys.sort} vs." \
-                                      " keys expected: #{expected_labels}"
+        begin
+          return labelset if keys_match?(labelset)
+        rescue ArgumentError
+          # If labelset contains keys that are a mixture of strings and symbols, this will
+          # raise when trying to sort them, but the error should be the same:
+          # InvalidLabelSetError
         end
 
-        labelset
+        raise InvalidLabelSetError, "labels must have the same signature " \
+                                    "(keys given: #{labelset.keys} vs." \
+                                    " keys expected: #{expected_labels}"
       end
 
       private